renamed bubble_chart to charting
